perf: remove dead work and vectorize loops in modifier construction
Part of the code review in #2706.

- shapesys: delete the never-read __shapesys_info tensor
- shapesys/staterror: hoist the mask astensor() out of the per-systematic
  reindex loop
- staterror: vectorize relerrs (np.where-guarded division, no scalar
  comprehension, no divide-by-zero warning)
- shapefactor: vectorize the access-field population per (s, t) with a
  clipped gather + np.where fallback
- normfactor: hoist param_viewer.get(pars) above the batch if/else
- shapesys: drop a duplicate get_backend() call in apply
- paramview: build all_indices with dtype="int" to match _all_indices

+++ src/pyhf/parameters/paramview.py
@@ -87,7 +87,7 @@ class ParamViewer:
     def _precompute(self):
         tensorlib, _ = get_backend()
 
-        self.all_indices = tensorlib.astensor(self._all_indices)
+        self.all_indices = tensorlib.astensor(self._all_indices, dtype="int")
         (
             self.index_selection,
             self.stitched,
